requirements.txt
apptools==5.2.0
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.8.2
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
comm==0.1.3
ConfigArgParse==1.5.3
configobj==5.0.8
contourpy==1.0.7
cycler==0.11.0
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
Django==4.2.1
django-annoying==0.10.6
envisage==7.0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
explainable-cnn==0.0.9
fastjsonschema==2.16.3
fitting==1.0.6
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.38.0
geomdl==5.0.0
greenlet==3.1.1
idna==3.6
imageio==2.35.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lazy_loader==0.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mayavi==4.8.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
open3d==0.17.0
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.0
pandas==2.0.3
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
plotly==5.14.1
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyface==8.0.0
Pygments==2.15.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyrsistent==0.19.3
PySide2==5.15.2.1
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.0.2
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
shiboken2==5.15.2.1
six==1.16.0
sklearn==0.0.post4
SQLAlchemy==2.0.38
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.2.2
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.7.10
tomli==2.0.1
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
traits==6.4.1
traitsui==7.4.3
trimesh==4.5.3
typing_extensions==4.12.2
tzdata==2023.3
ultralytics==8.1.29
urllib3==2.2.1
vtk==9.2.6
wcwidth==0.2.6
Werkzeug==3.0.6
widgetsnbextension==4.0.7
wincertstore==0.2
wxPython==4.2.1
zipp==3.15.0
